backend/requirements.txt
fastapi
uvicorn
pydantic
pydantic-settings
sqlalchemy
psycopg2-binary
python-jose[cryptography]
# 密码哈希直接使用 bcrypt，避免 passlib 与 bcrypt 4.1+ 的版本检测冲突
bcrypt>=4.0.0
alembic
python-multipart
